LayoutTests/fast/forms/datalist/range-snap-to-datalist-expected.txt

Tests if clicking close to a tick mark snaps to it.

PASS input.value is "0"
PASS parseInt(input.value, 10) < 500 is true
PASS input.value is "0"
PASS parseInt(input.value, 10) < 500 is true
PASS input.value is "0"
PASS parseInt(input.value, 10) < 500 is true
PASS input.value is "0"
PASS input.value is '500'
PASS input.value is "0"
PASS input.value is '500'
PASS input.value is "0"
PASS input.value is '500'
PASS input.value is "0"
PASS input.value is '500'
PASS input.value is "0"
PASS input.value is '500'
PASS input.value is "0"
PASS parseInt(input.value, 10) > 500 is true
PASS input.value is "0"
PASS parseInt(input.value, 10) > 500 is true
PASS input.value is "0"
PASS parseInt(input.value, 10) > 500 is true
PASS successfullyParsed is true

TEST COMPLETE
